smart-warehouse-project/flask-scanner-service/requirements.txt
# File: flask-scanner-service/requirements.txt

Flask==2.1.2
Werkzeug==2.3.8  # <-- CORRECTED VERSION
psycopg2-binary==2.9.3
gunicorn==20.1.0